% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_approx_fun.R
\name{create_approx_fun}
\alias{create_approx_fun}
\title{Title}
\usage{
create_approx_fun(
  f,
  from,
  to,
  N,
  K,
  language = "R",
  varname = "x",
  vectorize_JS = FALSE,
  pretty = TRUE
)
}
\arguments{
\item{f}{R function taking a single input}

\item{from}{Minimum x to evaluate}

\item{to}{Maximum x to evaluate}

\item{N}{Number of points to evaluate function at}

\item{K}{Number of knots to use when fitting the spline,
or a vector of knot points.}

\item{language}{Output type of the function. One of "R", "Rchar" for and R
function as a string, "JS" for JavaScript,}

\item{varname}{The name of the variable for the output function}

\item{vectorize_JS}{If the output is a JavaScript function, should code
be added so it is vectorized?}
}
\value{
A function or string
}
\description{
Title
}
\examples{

f1 <- function(x) {cos(exp(2*x^1.4))}
curve(f1, lwd=10)
f2 <- create_approx_fun(f1,0,1, 11, 4)
curve(f2, add=T, col=2, lwd=6, lty=2)
f3 <- create_approx_fun(f1,0,1, 51, 14)
curve(f3, add=T, col=3, lwd=6, lty=3)e
# Fewer points but concentrated in right end
f4 <- create_approx_fun(f1,0,1, 51, K=c(0,.3,.6,.7,.8,.85,.9,.95,.97,1))
curve(f4, add=T, col=4, lwd=6, lty=4)e

create_approx_fun(f1,0,1, 11, 4, language='rchar')
cat(create_approx_fun(f1,0,1, 11, 4, language='Js'))
cat(create_approx_fun(f1,0,1, 11, 4, language='Js', varname = "gamma"))
cat(create_approx_fun(f1,0,1, 11, 4, language='Js', vectorize_JS = T))
cat(create_approx_fun(f1,0,1, 11, 4, language='Js', vectorize_JS = T, varname="theta"))
cat(create_approx_fun(f1,0,1, 11, 4, language='Js'))
cat(create_approx_fun(f1,0,1, 11, 4, language='Js', pretty=FALSE))
}
